Refactor COCO inference test: replace `Model` with `RFDETRNano`, centralize dataset download logic in `conftest.py`
